futility: promote the single-owner TokenBucket beside its siblings
golf_hub's per-session stream bucket (#1240) earned a platform home the
moment a second consumer appeared: the same repo already carries a
per-key sliding window and a global thread-safe bucket, and the
single-owner injected-time shape is the deliberate third — no lock and
no clock call per admit, for budgets confined to one execution context.
The class and its fabricated-time tests move; golf_hub keeps only its
policy (RateLimits) and an alias. The README now tells the three shapes
apart and its stale //cpp target paths become the real ones.
